Added a GetById recursive method to EmployeeRepository
I did this because my previous method of using EF 'Include' and 'ThenInclude' was flawed. This assumed there would only ever be two tiers in the organization, when in fact it can be any amount. My new GetByIdRecursive method does use EF Include, but continues to traverse the tree to build the entire relationship, no matter how deep it is.